[APP-878] Created wrapper to hide getting installed apps from package manager
