+++ installed_apps/src/main/java/cm/aptoide/pt/installedapps/data/AptoideInstalledAppsRepository.kt
@@ -1,7 +1,5 @@
 package cm.aptoide.pt.installedapps.data
 
-import android.content.pm.PackageInfo
-import android.content.pm.PackageManager
 import cm.aptoide.pt.installedapps.data.database.LocalInstalledAppsRepository
 import cm.aptoide.pt.installedapps.data.database.model.InstalledAppEntity
 import cm.aptoide.pt.installedapps.domain.model.InstalledApp
@@ -15,32 +13,16 @@ import javax.inject.Singleton
 @Singleton
 class AptoideInstalledAppsRepository @Inject constructor(
   val localInstalledAppsRepository: LocalInstalledAppsRepository,
-  val packageManager: PackageManager
+  val installedAppsProvider: InstalledAppsProvider
 ) :
   InstalledAppsRepository {
 
   override suspend fun syncInstalledApps() {
     withContext(Dispatchers.IO) {
-      val installedAppsList = packageManager.getInstalledPackages(PackageManager.GET_META_DATA)
-      installedAppsList.sortWith { firstApp: PackageInfo, secondApp: PackageInfo -> ((firstApp.firstInstallTime - secondApp.firstInstallTime) / 1000).toInt() }
-      localInstalledAppsRepository.addListInstalledApps(mapInstalledAppsList(installedAppsList))
+      localInstalledAppsRepository.addListInstalledApps(installedAppsProvider.getInstalledApps())
     }
   }
 
-  private fun mapInstalledAppsList(installedAppsList: List<PackageInfo>): ArrayList<InstalledAppEntity> {
-    val installedList = ArrayList<InstalledAppEntity>()
-    for (packageInfo in installedAppsList) {
-      installedList.add(
-        InstalledAppEntity(
-          packageInfo.applicationInfo.loadLabel(packageManager).toString(),
-          packageInfo.packageName,
-          packageInfo.versionName,
-          "android.resource://" + packageInfo.packageName + "/" + packageInfo.applicationInfo.icon
-        )
-      )
-    }
-    return installedList
-  }
 
   override fun getInstalledApps(): Flow<List<InstalledApp>> {
     return localInstalledAppsRepository.getInstalledApps()
