+++ installed_apps/src/main/java/cm/aptoide/pt/installedapps/data/InstalledAppsProvider.kt
@@ -1,0 +1,8 @@
+package cm.aptoide.pt.installedapps.data
+
+import cm.aptoide.pt.installedapps.data.database.model.InstalledAppEntity
+
+interface InstalledAppsProvider {
+
+  fun getInstalledApps(): ArrayList<InstalledAppEntity>
+}
